chores: migrate to Next15, update and clean types
