+++ app/components/grid.tsx
@@ -1,65 +1,9 @@
 import Link from 'next/link';
 import React from 'react';
+import { type NeynarV1Cast, type User } from '../lib/types';
+import { renderCastText } from '../lib/utils';
 
-interface Cast {
-  hash: string;
-  parentHash: string | null;
-  parentUrl: string | null;
-  rootParentUrl: string | null;
-  threadHash: string;
-  parentAuthor: {
-    fid: number | null;
-  };
-  author: {
-    fid: number;
-    custodyAddress: string;
-    username: string;
-    displayName: string;
-    pfp: {
-      url: string;
-    };
-    profile: {
-      bio: {
-        text: string;
-        mentionedProfiles: any[];
-      };
-    };
-    followerCount: number;
-    followingCount: number;
-    verifications: string[];
-    verifiedAddresses: {
-      eth_addresses: string[];
-      sol_addresses: string[];
-    };
-    activeStatus: string;
-    powerBadge: boolean;
-  };
-  text: string;
-  timestamp: string;
-  embeds: {
-    url: string;
-  }[];
-  mentionedProfiles: any[];
-  reactions: {
-    count: number;
-    fids: number[];
-  };
-  recasts: {
-    count: number;
-    fids: number[];
-  };
-  recasters: any[];
-  replies: {
-    count: number;
-  };
-}
-
-interface ListRowProps {
-  cast: Cast;
-  isFirst: boolean;
-}
-
-const ListRow = ({ cast, isFirst }: ListRowProps) => {
+const ListRow = ({ cast, isFirst }: { cast: NeynarV1Cast, isFirst: boolean }) => {
   const rowRef = React.useRef<HTMLDivElement>(null);
 
   React.useEffect(() => {
@@ -103,10 +47,10 @@ const ListRow = ({ cast, isFirst }: ListRowProps) => {
     <div ref={rowRef} className={`border border-gray-300 p-4 flex flex-col justify-between h-full ${isFirst ? '' : 'border-t-0'}`}>
       <div>
         <div className="flex items-center mb-2">
-          <img src={cast.author.pfp.url} alt={`${cast.author.username}'s PFP`} className="w-6 h-6 rounded-full mr-2" />
-          <Link href={`https://warpcast.com/${cast.author.username}`} className="text-black font-medium">{cast.author.username}</Link>
+          <img src={(cast.author as any).avatar_url ?? ""} alt={`${(cast.author as any).fname ?? ""}'s PFP`} className="w-6 h-6 rounded-full mr-2" />
+          <Link href={`https://warpcast.com/${(cast.author as any).fname ?? ""}`} className="text-black font-medium">{(cast.author as any).fname ?? ""}</Link>
         </div>
-        <p className="text-gray-800 break-words">{cast.text}</p>
+        <p className="text-gray-800 break-words">{renderCastText(cast.text)}</p>
       </div>
       <div className="mt-2 flex justify-between items-end">
         <Link href={`/casts/${cast.hash}`} className="text-[#71579E] underline">link =&gt;</Link>
@@ -116,11 +60,7 @@ const ListRow = ({ cast, isFirst }: ListRowProps) => {
   );
 };
 
-interface ListProps {
-  casts: Cast[];
-}
-
-const Grid = ({ casts }: ListProps) => {
+const Grid = ({ casts }: { casts: NeynarV1Cast[]}) => {
   return (
     <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 border-l border-r border-gray-300">
       {casts.map((cast, index) => (
