+++ app/hooks/hub/use-hub-cast.tsx
@@ -1,45 +1,8 @@
+import { type HubCast } from '@/app/lib/types';
 import React from 'react';
 
-interface HubCastResponse {
-  data: {
-    type: string;
-    fid: number;
-    timestamp: number;
-    network: string;
-    author: {
-      username: string;
-      display_name: string;
-      pfp_url: string;
-      fid: number;
-    };
-    castAddBody: {
-      embedsDeprecated: any[];
-      mentions: any[];
-      parentUrl: string;
-      text: string;
-      mentionsPositions: any[];
-      embeds: {
-        url: string;
-      }[];
-      type: string;
-    };
-    replies: {
-      count: number;
-    };
-    reactions: {
-      likes_count: number;
-      recasts_count: number;
-    };
-  };
-  hash: string;
-  hashScheme: string;
-  signature: string;
-  signatureScheme: string;
-  signer: string;
-}
-
 const useHubCast = (fid: number | null, hash: string, type: 'neynar' | 'warpcast') => {
-  const [cast, setCast] = React.useState<HubCastResponse | null>(null);
+  const [cast, setCast] = React.useState<HubCast | null>(null);
   const [loading, setLoading] = React.useState(true);
   const [error, setError] = React.useState<string | null>(null);
