+++ app/hooks/neynar/use-neynar-cast.tsx
@@ -1,84 +1,8 @@
+import { NeynarV2Cast } from '@/app/lib/types';
 import React from 'react';
 
-interface NeynarCastResponse {
-  object: string;
-  hash: string;
-  thread_hash: string;
-  parent_hash: string | null;
-  parent_url: string | null;
-  root_parent_url: string | null;
-  parent_author: {
-    fid: number | null;
-  };
-  author: {
-    object: string;
-    fid: number;
-    custody_address: string;
-    username: string;
-    display_name: string;
-    pfp_url: string;
-    profile: {
-      bio: {
-        text: string;
-      };
-    };
-    follower_count: number;
-    following_count: number;
-    verifications: string[];
-    verified_addresses: {
-      eth_addresses: string[];
-      sol_addresses: string[];
-    };
-    active_status: string;
-    power_badge: boolean;
-  };
-  text: string;
-  timestamp: string;
-  embeds: {
-    url: string;
-  }[];
-  reactions: {
-    likes_count: number;
-    recasts_count: number;
-    likes: {
-      fid: number;
-      fname: string;
-    }[];
-    recasts: {
-      fid: number;
-      fname: string;
-    }[];
-  };
-  replies: {
-    count: number;
-  };
-  mentioned_profiles: {
-    object: string;
-    fid: number;
-    custody_address: string;
-    username: string;
-    display_name: string;
-    pfp_url: string;
-    profile: {
-      bio: {
-        text: string;
-        mentioned_profiles: any[];
-      };
-    };
-    follower_count: number;
-    following_count: number;
-    verifications: string[];
-    verified_addresses: {
-      eth_addresses: string[];
-      sol_addresses: string[];
-    };
-    active_status: string;
-    power_badge: boolean;
-  }[];
-}
-
 const useNeynarCast = (identifier: string, type: 'url' | 'hash') => {
-  const [cast, setCast] = React.useState<NeynarCastResponse | null>(null);
+  const [cast, setCast] = React.useState<NeynarV2Cast | null>(null);
   const [loading, setLoading] = React.useState(true);
   const [error, setError] = React.useState<string | null>(null);
