+++ middleware.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
-import { BASE_URL } from './app/lib/consts';
+import { BASE_URL } from './app/lib/utils';
 
 export function middleware(request: NextRequest) {
   const origin = request.nextUrl.origin || request.headers.get('origin') || request.headers.get('referer');
