+++ next.config.js
@@ -1,13 +1,5 @@
-// next.config.js
-
-const withPWA = require('next-pwa')({
-    dest: 'public'
-  })
-  
-  module.exports = withPWA({
+module.exports = {
     eslint: {
-      // Warning: This allows production builds to successfully complete even if
-      // your project has ESLint errors.
       ignoreDuringBuilds: true,
     },
     images: {
@@ -107,4 +99,4 @@ const withPWA = require('next-pwa')({
         },
       ],
     },
-  });
+  };
